implement OpenMetrics support for real
I discovered that thanos-federate-proxy lies about its capabilities:
it's not truly OpenMetrics compliant.

The prometheus metrics scrape endpoint representation differs from
OpenMetrics in subtle ways. Consumers who only support OpenMetrics
fail to scrape anything produced by `thanos-federate-proxy`.

For context, `thanos-federate-proxy` retrieves from the upstream
using Prometheus Query API. This API returns `model.Value`, a wire
representation that does not have all the same information presented
in a scrape endpoint (`MetricsFamily`). If we transform it to `MetricsFamily`
_somehow_, we could use the `expfmt` package from `prometheus/common`
library with the official bits for content negotiation and serialization
in the scrape endpoint textual representation.

So that's what this PR does.

The challenge here was how to transform `model.Value` to `MetricsFamily`.
I discovered Prometheus offers `Metadata API`, which gives us the missing bits.
Unfortunately, Thanos does not expose this API, so this commit requires you
to wire the metadata API through the upstream prometheus. API Queries will
hit Thanos, and metadata API calls will hit prometheus. The latter is done
outside the critical path on an interval currently hardcoded to 1 minute.
There could be a scenario where Metadata for a new metric is not yet
present (as the async worker hasn't refreshed the metadata) and in that
case the metric will be set to `unknown` type.

I don't know if the transformation is done correctly, but so far, it
looks to deliver what we need. Please be sure to proceed with caution,
this is experimental.