feat: introduce capability to change log verbosity
A customer's Loki is ingesting millions of log lines a day from
thanos-federate-proxy due to an inherent many-to-one mapping in
thanos-federate-proxy, which causes problems when metric families change
their declared type or documentation string. More details about how and
why metadata for a metric family changes can be found here:
https://linear.app/authzed/issue/OSS-299/fix-issue-with-thanos-federate-proxy-and-multiple-metadata-entries.

This many-to-one problem is an inherent issue in thanos-federate-proxy,
so to work around the issue of excessive logging, this commit exposes
klogs flags for controlling logging verbosity and sets the level of the
multiple-metadata warning at a level above the default verbosity of 0.

Signed-off-by: squat <lserven@gmail.com>